requirements.txt
arrow==1.2.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Deprecated==1.2.13
Django==3.2.7
django-ckeditor==6.1.0
django-crontab==0.7.1
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-redis==5.0.0
djangorestframework==3.12.4
idna==3.3
Jinja2==3.0.2
jinja2-time==0.2.0
kafka-python==2.0.2
make==0.1.6.post2
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-gettext==4.0
pytz==2021.3
redis==4.0.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
wrapt==1.13.3
